pyproject.toml
[tool.poetry]
name = "wordle-csit-pro"
version = "0.1.0"
description = ""
authors = ["Nikita B.<maxim.rossa@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.7"
fastapi = "^0.114.1"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
cryptography = "^43.0.1"
password-strength = "^0.0.3.post2"

[tool.poetry.scripts]
app = "src.app:app"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
ruff = "0.5.5"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
	"ALL",
]

ignore = [
	"Q000", # single quotes
]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"